allow all font-size values in config closes #32
Replaced the dropdown choices with a custom text box. By default it has 
12px as per Stylus default, this is more flexible because it will 
support all spec values like 
https://developer.mozilla.org/en-US/docs/Web/CSS/font-size describes.

On my original test for this change, I found it may break backwards 
compatibility because the custom entry was the old value `Large 14px` 
for me and `Large 14px` is invalid, so if affected change this to a 
valid value.

So I'll add this, with this caution and will bump the minor version, to 
reflect this possibility.